ComputerStorePOS/src/test/java/com/aiden/computerstorepos/test/MemoryTest.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */

package com.aiden.computerstorepos.test;

import com.aiden.computerstorepos.conf.AppConfig;
import com.aiden.computerstorepos.domain.Memory;
import com.aiden.computerstorepos.services.MemoryService;
import org.springframework.context.ApplicationContext;
import org.springframework.context.annotation.AnnotationConfigApplicationContext;
import org.testng.Assert;
import static org.testng.Assert.*;
import org.testng.annotations.AfterClass;
import org.testng.annotations.AfterMethod;
import org.testng.annotations.BeforeClass;
import org.testng.annotations.BeforeMethod;
import org.testng.annotations.Test;

/**
 *
 * @author Aidem
 */
public class MemoryTest {
    private MemoryService service;
    public MemoryTest() {
    }

  @Test
    public void testMemory() throws Exception {
        Memory data = service.getMemory();
        Assert.assertEquals(data.getProductNumber(),"HX318C10FWK2/8");
    }
    @BeforeClass
    public static void setUpClass() throws Exception {
    }

    @AfterClass
    public static void tearDownClass() throws Exception {
    }

    @BeforeMethod
    public void setUpMethod() throws Exception {
        ApplicationContext ctx = new AnnotationConfigApplicationContext(AppConfig.class);
        service = (MemoryService)ctx.getBean("Memory");
    }

    @AfterMethod
    public void tearDownMethod() throws Exception {
    }
}
